refactor: add a deskulpt-common crate
