+++ Cargo.toml
@@ -42,6 +42,7 @@ tokio                          = "1.47.1"
 uuid                           = "1.16.0"
 
 # Deskulpt crates
+deskulpt-common        = { version = "0.0.1", path = "crates/deskulpt-common" }
 deskulpt-core          = { version = "0.0.1", path = "crates/deskulpt-core" }
 deskulpt-macros        = { version = "0.0.1", path = "crates/deskulpt-macros" }
 deskulpt-plugin        = { version = "0.0.1", path = "crates/deskulpt-plugin" }
@@ -52,9 +53,6 @@ deskulpt-workspace     = { version = "0.0.1", path = "crates/deskulpt-workspace"
 deskulpt-plugin-fs  = { version = "0.0.1", path = "crates/deskulpt-plugin-fs" }
 deskulpt-plugin-sys = { version = "0.0.1", path = "crates/deskulpt-plugin-sys" }
 sysinfo             = "0.33.1"
-
-[workspace.metadata.cargo-shear]
-ignored = ["deskulpt-macros"] # Remove once macros crate is applied
 
 [profile.dev]
 incremental = true
