+++ crates/deskulpt-common/README.md
@@ -1,0 +1,3 @@
+This crate implements shared utilities for the [Deskulpt](https://csci-shu-410-se-project.github.io/) application.
+
+⚠️ This crate is meant to be used internally by the Deskulpt application and is not designed to support plugin authors or other users directly. Private items are documented for reference of Deskulpt developers.
