+++ crates/deskulpt-core/Cargo.toml
@@ -11,7 +11,7 @@ version    = { workspace = true }
 
 [dependencies]
 anyhow                       = { workspace = true }
-deskulpt-macros              = { workspace = true }
+deskulpt-common              = { workspace = true }
 dunce                        = { workspace = true }
 either                       = { workspace = true }
 once_cell                    = { workspace = true }
