+++ crates/deskulpt-core/src/events.rs
@@ -1,40 +1,9 @@
 //! Deskulpt core events.
 
-use anyhow::Result;
-use deskulpt_macros::Event;
+use deskulpt_common::event::Event;
 use serde::{Deserialize, Serialize};
-use tauri::{Emitter, Runtime};
 
 use crate::settings::{Theme, WidgetSettings};
-use crate::window::DeskulptWindow;
-
-/// Trait for Deskulpt events.
-///
-/// This trait should be derived with [`Event`](deskulpt_macros::Event).
-pub trait Event: specta::Type + Serialize {
-    /// The name of the event.
-    const NAME: &'static str;
-
-    /// Emit the event to all target.
-    fn emit<R, E>(&self, emitter: &E) -> Result<()>
-    where
-        R: Runtime,
-        E: Emitter<R>,
-    {
-        emitter.emit(Self::NAME, self)?;
-        Ok(())
-    }
-
-    /// Emit the event to the specified window.
-    fn emit_to<R, E>(&self, emitter: &E, window: DeskulptWindow) -> Result<()>
-    where
-        R: Runtime,
-        E: Emitter<R>,
-    {
-        emitter.emit_to(window, Self::NAME, self)?;
-        Ok(())
-    }
-}
 
 /// Event for exiting the application.
 ///
