+++ crates/deskulpt-macros/README.md
@@ -1,3 +1,3 @@
 This crate implements the procedural macros used by the [Deskulpt](https://csci-shu-410-se-project.github.io/) application.
 
-⚠️ This crate is meant to be used internally by the Deskulpt application and is not designed to support plugin authors or other users directly. Private items are documented for reference of Deskulpt developers.
+⚠️ This crate is not meant to be used directly; use corresponding items from other Deskulpt crates instead. Private items are documented for reference of Deskulpt developers.
