+++ crates/deskulpt-plugin-macros/README.md
@@ -1,3 +1,3 @@
-This crate implements procedural macros for the `deskulpt-plugin` crate.
+This crate implements the procedural macros for the `deskulpt-plugin` crate.
 
 ⚠️ This crate is not meant to be used directly; use corresponding items from `deskulpt-plugin` instead. Private items are documented for reference of Deskulpt developers.
